+++ frontend/src/pages/Invite/ScanTheirNICEID.js
@@ -1,22 +1,11 @@
 import React, { useState } from 'react'
-import { QRReader } from '@digicatapult/ui-component-library'
 
-import {
-  SmallThinText,
-  Button,
-  QRReaderWrapper,
-  ContentWrapper,
-} from '../../components/shared'
+import { SmallThinText, Button, ContentWrapper } from '../../components/shared'
 import WrapperWithHeader from '../../components/Header'
+import QRReaderSection from '../../components/QRCode'
 
 const ScanTheirNiceId = () => {
   const [hasScanned, setHasScanned] = useState(false)
-
-  const handleQrCode = (qr) => {
-    // eslint-disable-next-line
-    console.log(`Scanned: ${qr}`)
-    setHasScanned(true)
-  }
 
   const handleBackToProfile = () => {
     window.location.href = `/ssi-profile`
@@ -26,21 +15,11 @@ const ScanTheirNiceId = () => {
     return (
       <>
         <h2>Add Company to your NICE Network</h2>
-        <SmallThinText>
-          If their company is already on NICE network, simply scan their NICE ID
-          QR Code to add them as a NICE connection.
-        </SmallThinText>
-        <QRReaderWrapper>
-          <QRReader
-            viewFinderColor="#fff"
-            viewFinderVariant="viewfinder-cross-med"
-            onResult={handleQrCode}
-            mirror={true}
-          />
-        </QRReaderWrapper>
-        <SmallThinText>
-          Please enable the camera access to scan the QR codes.
-        </SmallThinText>
+        <QRReaderSection
+          instructionText={`If their company is already on NICE network, simply scan their NICE ID
+          QR Code to add them as a NICE connection.`}
+          setHasScanned={setHasScanned}
+        ></QRReaderSection>
         <Button onClick={handleBackToProfile}>{'<'}</Button>
       </>
     )
